Knockback with new weapon system

+++ Assets/_Scripts/Characters/ICharacterHealth.cs
@@ -2,6 +2,6 @@ public interface ICharacterHealth
 {
     public void TakeDamage(int amount);
     public void Die();
-    public int GetCurrentHealth();
-    public int GetMaxHealth();
+    public float GetCurrentHealth();
+    public float GetMaxHealth();
 }
